Enhance channel configuration for JWT impl
