+++ server/channel/channel_atom.go
@@ -1,30 +1,45 @@
 package channel
 
 import (
+	"context"
 	"fmt"
 
 	"golang.org/x/xerrors"
 
 	"github.com/saiya/dsps/server/config"
 	"github.com/saiya/dsps/server/domain"
+	jwtv "github.com/saiya/dsps/server/jwt/validator"
 )
 
 // channelAtom is an Channel implementation corresponds to a ChannelConfiguration
 type channelAtom struct {
 	// ChannelConfig that this object corresponds to.
 	config *config.ChannelConfig
+
+	JwtValidator jwtv.Validator
 }
 
-func newChannelAtom(config *config.ChannelConfig, validate bool) (*channelAtom, error) {
+func newChannelAtom(ctx context.Context, config *config.ChannelConfig, clock domain.SystemClock, validate bool) (*channelAtom, error) {
 	atom := &channelAtom{
 		config: config,
 	}
 	if validate {
 		if err := atom.validate(); err != nil {
-			return &channelAtom{}, err
+			return nil, err
 		}
 	}
+
+	jwtValidator, err := jwtv.NewValidator(ctx, config.Jwt, clock)
+	if err != nil {
+		return nil, err
+	}
+	atom.JwtValidator = jwtValidator
+
 	return atom, nil
+}
+
+func (c *channelAtom) String() string {
+	return c.config.Regex.String()
 }
 
 func (c *channelAtom) validate() error {
@@ -35,13 +50,6 @@ func (c *channelAtom) validate() error {
 }
 
 func (c *channelAtom) validateTemplateStrings() error {
-	dummy := make(map[string]interface{})
-	dummyRegexMatches := make(map[string]string)
-	for _, name := range c.config.Regex.GroupNames() {
-		dummyRegexMatches[name] = "dummy"
-	}
-	dummy["regex"] = dummyRegexMatches
-
 	templates := make(map[string]domain.TemplateString)
 	for i, webhook := range c.config.Webhooks {
 		templates[fmt.Sprintf("webhooks[%d].url", i)] = *webhook.URL
@@ -55,12 +63,33 @@ func (c *channelAtom) validateTemplateStrings() error {
 		}
 	}
 
+	dummy := c.dummyTemplateEnvironment()
 	for path, tpl := range templates {
 		if _, err := tpl.Execute(dummy); err != nil {
 			return xerrors.Errorf("invalid template found on %s: %w", path, err)
 		}
 	}
 	return nil
+}
+
+func (c *channelAtom) TemplateEnvironmentOf(id domain.ChannelID) domain.TemplateStringEnv {
+	matches := c.config.Regex.Match(true, string(id))
+	if matches == nil {
+		return nil
+	}
+	return map[string]interface{}{
+		"regex": matches,
+	}
+}
+
+func (c *channelAtom) dummyTemplateEnvironment() domain.TemplateStringEnv {
+	dummyRegexMatches := make(map[string]string)
+	for _, name := range c.config.Regex.GroupNames() {
+		dummyRegexMatches[name] = "dummy"
+	}
+	return map[string]interface{}{
+		"regex": dummyRegexMatches,
+	}
 }
 
 func (c *channelAtom) IsMatch(id domain.ChannelID) bool {
