+++ server/jwt/alg.go
@@ -1,0 +1,11 @@
+package jwt
+
+import (
+	"github.com/saiya/dsps/server/domain"
+)
+
+// ValidateAlg validates JWT alg
+func ValidateAlg(alg domain.JwtAlg) error {
+	// TODO: Implement
+	return nil
+}
